fix: resolve ambiguous trigonometric function calls
- Add Foundation import for cos/sin functions
- Add explicit CGFloat casting for trigonometric results
- Fix compilation errors in AnimatedProfileMetric and ProfileCompletionCelebration

+++ Gym_API/Views/Components/Profile/AnimatedProfileMetric.swift
@@ -1,4 +1,5 @@
 import SwiftUI
+import Foundation
 
 // MARK: - Animated Profile Metric
 struct AnimatedProfileMetric: View {
@@ -187,8 +188,8 @@ struct AnimatedProfileMetric: View {
         let angle = Double(index) * 60.0 * .pi / 180.0
         let distance: CGFloat = 25
         return CGSize(
-            width: cos(angle) * distance,
-            height: sin(angle) * distance
+            width: CGFloat(cos(angle)) * distance,
+            height: CGFloat(sin(angle)) * distance
         )
     }
 }
